Extract function `getSubscriptionInfo` from `Attachments` to `utils`
