+++ react/utils/index.tsx
@@ -21,3 +21,63 @@ export interface PaymentGroupInfo {
   url: string | null
   value: number
 }
+
+export const getSubscriptionInfo = (
+  attachmentItem: Attachment,
+  intl: ReactIntl.InjectedIntl
+) => {
+  const isSubscription = attachmentItem.name.indexOf('vtex.subscription') === 0
+  if (!isSubscription) {
+    return { isSubscription: false }
+  }
+  const vtexSubsPrefix = 'vtex.subscription.key.'
+  const subsFrequency: string =
+    attachmentItem.content[`${vtexSubsPrefix}frequency`]
+  const subsPurchaseDay =
+    attachmentItem.content[`${vtexSubsPrefix}purchaseday`]
+  const subsValidityBegin =
+    attachmentItem.content[`${vtexSubsPrefix}validity.begin`]
+  const subsValidityEnd =
+    attachmentItem.content[`${vtexSubsPrefix}validity.end`]
+
+  // This matches a key in the format: "3 months"
+  const numberPeriodRegex = /([\d]{1,3}) ([A-z]{1,20})/
+  // This matches a key in the format: "weekly"
+  const wordlyPeriodRegex = /([A-z]{1,20})/
+
+  const subsFrequencyString = numberPeriodRegex.test(subsFrequency)
+    ? intl.formatMessage(
+      {
+        id: `items.attachments.subscription.frequency.${
+          (numberPeriodRegex.exec(subsFrequency) || [])[2]
+          }`,
+      },
+      {
+        frequencyNumber: parseInt(
+          (numberPeriodRegex.exec(subsFrequency) || [])[1],
+          10
+        ),
+      }
+    )
+    : intl.formatMessage({
+      id: `items.attachments.subscription.frequency.${
+        (wordlyPeriodRegex.exec(subsFrequency) || [])[1]
+        }`,
+    })
+
+  const subsPurchaseDayString =
+    subsPurchaseDay !== ''
+      ? intl.formatMessage(
+        { id: 'items.attachments.subscription.purchaseday' },
+        { purchaseday: subsPurchaseDay }
+      )
+      : null
+
+  return {
+    isSubscription: true,
+    subsFrequency: subsFrequencyString,
+    subsPurchaseDay: subsPurchaseDayString,
+    subsValidityBegin,
+    subsValidityEnd,
+  }
+}
